feat: refactor UI Component file
