+++ src/App.jsx
@@ -1,10 +1,8 @@
 import './App.scss'
 
-import UserFunctionTab from './UIComponents/tab/UserFunctionTab'
 function App() {
   return (
     <div className="App">
-      <UserFunctionTab />
     </div>
   )
 }
